% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnoms.R
\name{plotAnoms}
\alias{plotAnoms}
\title{Plots streamflow anomalies}
\usage{
plotAnoms(data, ...)
}
\arguments{
\item{data}{is the anomaly list from the function 
\link{compAnom}.}

\item{...}{further arguments to be passed to plotting method (see \link{par}).}
}
\value{
a plot.
}
\description{
Function to plot hydrologic time series and anomalies
}
\examples{
 \dontrun{
q05054000.85 <- importDVs("05054000", sdate="1985-01-01", edate="2010-09-30")
anoms05054000 <- compAnom(q05054000.85, which=1)
plotAnoms(anoms05054000)
 }
}
\keyword{hplot}
\keyword{multivariate}
\keyword{ts}

