% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tellMeSiteURL}
\alias{tellMeSiteURL}
\title{USGS Site Information Service URL}
\usage{
tellMeSiteURL(staid)
}
\arguments{
\item{staid}{is the USGS site identification number, which
is usually eight digits long, but can be longer.  Users may search for
surface-water sites and obtain station identification numbers using the
USGS Site Web Service,
\url{http://waterservices.usgs.gov/rest/Site-Service.html} (U.S. Geological
Survey, 2012b); using the National Water Information System: Mapper,
\url{http://wdr.water.usgs.gov/nwisgmap/} (U.S. Geological Survey, 2012a);
or using the National Water Information System: Web Interface to daily
surface-water data,
\url{http://waterdata.usgs.gov/nwis/dv/?referred_module=sw} (U.S.
Geological Survey, 2012c).  The site identification number needs to be
entered as a character, that is in quotes, because many USGS streamgage
numbers begin with zero and the leading zero is necessary.}
}
\value{
URL for USGS site information
}
\description{
Function that returns USGS Site Information Service URL for troubleshooting or
building a URL for other purposes.
}
\examples{
tellMeSiteURL("05054000")
}
\references{
U.S. Geological Survey, 2012a, National Water Information System: Mapper,
accessed September 7, 2012, at
\url{http://wdr.water.usgs.gov/nwisgmap/}.

U.S. Geological Survey, 2012b, USGS site web service:
REST Web Services, accessed September 7, 2012, at
\url{http://waterservices.usgs.gov/rest/Site-Service.html}.

U.S. Geological Survey, 2012c, USGS surface-water daily data for the Nation:
National Water Information System: Web Interface, accessed September 7,
2012, at \url{http://waterdata.usgs.gov/nwis/dv/?referred_module=sw}.
}
\keyword{utilities}

